+++ src/appleseed/renderer/kernel/rendering/debug/debugsamplerenderer.cpp
@@ -74,11 +74,11 @@ namespace
         void render_sample(
             SamplingContext&            sampling_context,
             const PixelContext&         pixel_context,
-            const Vector2d&             image_point,
+            const Vector2f&             image_point,
             AOVAccumulatorContainer&    aov_accumulators,
             ShadingResult&              shading_result) override
         {
-            const Vector2d v = Vector2d(0.5) - image_point;
+            const Vector2f v = Vector2f(0.5) - image_point;
             const double d = norm(v) * 2.0;
             const double M = 0.01;
             const float c =
